requirements.txt
evernote3==1.25.12
python-decouple==3.1
